Fix shadow issue
Shadow applied to the header element was overflowing onto the sidebar
div element. Resolved the issue by creating a new div element to contain
the header and main elements separately from the sidebar div element and
set the overflow property to hidden for this container. Also used the
main element's background color as the container's background color and
made the main element transparent so that the header's shadow is
visible.